feat: add check if notifications are enabled before sending server-side
